Refactor communication between core and extensions
In the previous implementation, the extensions were using only one way
to communicate with each other, which was getting a notification when a
config value was set or changed. This was problematic on startup because
there was no order in which the extensions were activated and when
messages were handled. b5bf26ac084823758be18453bb61579fefceb57c
partially fixed this issue by adding a way to wait for `qt-cpp` to be
ready, but it was still not enough. The extension should read config
values when they are ready, not when messages are sent from other
extensions.

Remove the `lazy` initialization parameter in `Project` classes. Instead
of using that parameter, we remove the firing event from `addProject()`
and only fire when the project is added after startup. This way, we
don't need the `lazy` parameter anymore.

Since `processMessage()` is not used during startup, this commit also
fixes QTBUG-131702 on the extension side.

qt-lib:

This commit changes `QtWorkspaceConfigMessage` and removes values in it.
`QtWorkspaceConfigMessage` is now used only to notify that a value or
values were set. The value `CoreApi` can be accessed by `getValue`
and `setValue`.

* Update the `CoreApi` interface
* Add `setValue`
* Rename `update` to `notify`
* Remove `get<T>()` from `QtWorkspaceConfigMessage`

qt-core:

* Remove the internal checking mechanism to understand if a value was
set or changed. Instead, just store values.

qt-cpp:

* Update `processMessage()` for the new usage.
* Initialize config values explicitly during startup.

qt-qml:

* Update `processMessage()` for the new usage.
* Separate starting `qmlls` from the constructor.
 It should be started when the conditions are met.
* Get config values and update parameters inside `onProjectAdded()`

Task-number: QTBUG-131702
Change-Id: If9831ea1257d123f777e6ae2afb92f33942dd3da
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>